Push a variable back to an int to deal with signed vs unsigned comparison warnings
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1974 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/RenderMain/RenderSortPoly.cpp
@@ -312,7 +312,7 @@ clipping_window_data *RenderSortPolyClass::build_clipping_windows(
 				if (j!=NONE) /* if the endpoint was not a duplicate */
 				{
 					/* expand the array, if necessary, and add the new endpoint */
-					size_t Length = AccumulatedEndpointClips.size();
+					int Length = AccumulatedEndpointClips.size();
 					AccumulatedEndpointClips.push_back(NULL);
 					assert(AccumulatedEndpointClips.size() <= 32767);		// Originally a short value
 					if (j!=Length) memmove(&AccumulatedEndpointClips[j+1], &AccumulatedEndpointClips[j],
